move class ver. but partial initialization
