+++ test/test_projection_class.py
@@ -1,9 +1,8 @@
 from pathlib import Path
 import numpy as np
 
-from disparity_view.o3d_project import gen_tvec, DEPTH_SCALE
+from disparity_view.o3d_project import gen_tvec, DEPTH_SCALE, StereoCamera
 from disparity_view.o3d_project import as_extrinsics
-from disparity_view.projection_class import StereoCamera
 from disparity_view.util import safer_imsave
 import skimage.io
